+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Improvements
 
+* `ETS()` now supports missing values.
 * Documentation improvements.
 
 ## Bug fixes